requirements/development.txt
-r base.txt
requests_mock
model_bakery
pytest
pytest-cov
pytest-django
ipdb
ipython
django-cors-headers
